+++ src/index.ts
@@ -481,6 +481,8 @@ export {
 	defineExtensionDecoder,
 	defineExtensionDecoderMap,
 	findExtension,
+	getSubjectPublicKey,
+	getSubjectPublicKeyOrThrow,
 	parseCertificateChainPem,
 	parseCertificateDer,
 	parseCertificateDerOrThrow,